+++ Web/vue3/src/TUIRoom/components/RoomContent/LocalScreenStream/index.vue
@@ -5,15 +5,14 @@
         <svg-icon :icon="ScreenSharingIcon"></svg-icon>
         <span class="text">{{ t('You are sharing the screen...') }}</span>
       </div>
-      <Button size="default" class="stop-button" @click="openStopConfirmDialog">
+      <tui-button size="default" class="stop-button" @click="openStopConfirmDialog">
         {{ t('End sharing') }}
-      </Button>
+      </tui-button>
       <Dialog
-        :model-value="showStopShareRegion"
+        v-model="showStopShareRegion"
         width="420px"
         :title="t('End sharing')"
         :modal="true"
-        :before-close="closeScreenShareDialog"
         :close-on-click-modal="true"
         :append-to-room-container="true"
       >
@@ -21,8 +20,8 @@
           {{ t('Others will no longer see your screen after you stop sharing. Are you sure you want to stop?') }}</span>
         <template #footer>
           <span>
-            <Button class="dialog-button" size="default" @click="stopScreenSharing">{{ t('End sharing') }}</Button>
-            <Button type="primary" size="default" @click="showStopShareRegion = false">{{ t('Cancel') }}</Button>
+            <tui-button class="dialog-button" size="default" @click="stopScreenSharing">{{ t('End sharing') }}</tui-button>
+            <tui-button type="primary" size="default" @click="showStopShareRegion = false">{{ t('Cancel') }}</tui-button>
           </span>
         </template>
       </Dialog>
@@ -34,7 +33,7 @@
 import { ref } from 'vue';
 import SvgIcon from '../../common/base/SvgIcon.vue';
 import ScreenSharingIcon from '../../common/icons/ScreenSharingIcon.vue';
-import Button from '../../common/base/Button.vue';
+import TuiButton from '../../common/base/Button.vue';
 import Dialog from '../../common/base/Dialog';
 import eventBus from '../../../hooks/useMitt';
 import { useI18n } from '../../../locales';
@@ -54,10 +53,6 @@ function openStopConfirmDialog() {
 function stopScreenSharing() {
   showStopShareRegion.value = false;
   eventBus.emit('ScreenShare:stopScreenShare');
-}
-
-function closeScreenShareDialog() {
-  showStopShareRegion.value = false;
 }
 </script>